requirements.txt
bleak
python-dotenv
